pyproject.toml
[project]
name = "intelligentdocref"
version = "0.1.0"
description = "Local-first Intelligent Document Reference"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.128.3",
    "google-genai>=1.0.0",
    "tiktoken>=0.8.0",
    "httpx>=0.28.1",
    "numpy>=2.4.2",
    "openai>=2.17.0",
    "pillow>=12.1.0",
    "pymupdf>=1.25.0",
    "python-dotenv>=1.2.1",
    "sqlite-vec>=0.1.6",
    "typer>=0.21.1",
    "uvicorn>=0.40.0",
    "voyageai>=0.3.7",
    "watchdog>=6.0.0",
    "pytesseract>=0.3.10",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.10"]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]

[tool.isort]
profile = "black"
line_length = 88
known_first_party = [
    "backend",
    "config",
    "core",
    "db",
    "embeddings",
    "evaluation",
    "inference",
    "ingestion",
    "jobs",
    "model_clients",
    "watcher",
]

[tool.ruff]
line-length = 88
target-version = "py311"
